Move code into dedicated file to reduce size
