+++ tests/constants.py
@@ -175,7 +175,6 @@ JULIA = "julia"
 GPU_TRITON = "triton"
 GPU_NIM = "nim"
 GPU_NIM_SIDECAR = "nim_sidecar"
-GPU_NIM_EMBEDQA = "nim_embedqa"
 GPU_VLLM = "vllm"
 PYTHON_TEXT_GENERATION = "python_text_generation"
 PYTHON_GEO_POINT = "python_geo_point"
