+++ tests/functional/run_integration_tests_in_framework_container.sh
@@ -29,7 +29,8 @@ pip install pytest pytest-xdist
 TESTS_TO_RUN="tests/functional/test_inference_per_framework.py \
               tests/functional/test_fit_per_framework.py \
               tests/functional/test_other_cases_per_framework.py \
-              tests/functional/test_unstructured_mode_per_framework.py
+              tests/functional/test_unstructured_mode_per_framework.py \
+              tests/functional/test_inference_gpu_predictors.py \
              "
 
 # install 'pack' package in R env for tests
